DB Connection parameters are transferred to the resource file. Rename MySQLConnector to DBConnector.
