--- src/main/resources/database.properties
@@ -1,4 +1,0 @@
-url=jdbc:mysql://localhost:3306/hr_app_db?autoReconnect=true&useSSL=false
-login=root
-password=root
-driver=com.mysql.cj.jdbc.Driver
