+++ src/main/resources/db/db.properties
@@ -1,0 +1,4 @@
+db.url=jdbc:mysql://localhost:3306/hr_app_db?autoReconnect=true&useSSL=false
+db.username=root
+db.password=root
+db.driver=com.mysql.cj.jdbc.Driver
